+++ hslua-core/src/HsLua/Core/Primary.hs
@@ -782,6 +782,21 @@ replace :: StackIndex -> LuaE e ()
 replace n = liftLua $ \l ->  lua_replace l n
 {-# INLINABLE replace #-}
 
+-- | Rotates the stack elements between the valid index @idx@ and the
+-- top of the stack. The elements are rotated @n@ positions in the
+-- direction of the top, for a positive @n@, or @-n@ positions in the
+-- direction of the bottom, for a negative @n@. The absolute value of
+-- @n@ must not be greater than the size of the slice being rotated.
+-- This function cannot be called with a pseudo-index, because a
+-- pseudo-index is not an actual stack position.
+--
+-- <https://www.lua.org/manual/5.3/manual.html#lua_rotate>
+rotate :: StackIndex  -- ^ @idx@
+       -> Int         -- ^ @n@
+       -> LuaE e ()
+rotate idx n = liftLua $ \l -> lua_rotate l idx (fromIntegral n)
+{-# INLINABLE rotate #-}
+
 -- | Does the equivalent to @t[k] = v@, where @t@ is the value at the
 -- given index and @v@ is the value at the top of the stack.
 --